% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_edge_dist_attrib.R
\name{add_edge_dist_attrib}
\alias{add_edge_dist_attrib}
\title{Validation Function}
\usage{
add_edge_dist_attrib(graph)
}
\arguments{
\item{graph}{igraph object with V(g)$X and V(g)$Y having the X and Y co-ordinates of each node}
}
\description{
This function adds a dist attribute to edges as a measure of the Euclidean distance of the connection
}
\keyword{:}
\keyword{attributes}
\keyword{check}

