% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{cat_function}
\alias{cat_function}
\title{A trial Function}
\usage{
cat_function()
}
\arguments{
\item{:}{None}
}
\description{
This function allows you to express your love of Hello World
}
\examples{
trial
}
\keyword{:}
\keyword{None}

