% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_connection_distance_distribution.R
\name{check_connection_distance_distribution}
\alias{check_connection_distance_distribution}
\title{A trial Function}
\usage{
check_connection_distance_distribution(graph)
}
\arguments{
\item{:}{igraph object with V(g)$X and V(g)$Y having the X and Y co-ordinates of each node}
}
\description{
This function allows you to express your love of Hello World
}
\examples{
check_connection_distance_distribution()
}
\keyword{:}
\keyword{clustering}
\keyword{spatial}

