% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spatial_attribs.R
\name{check_spatial_attribs}
\alias{check_spatial_attribs}
\title{Validation Function}
\usage{
check_spatial_attribs(graph)
}
\arguments{
\item{:}{igraph object with V(g)$X and V(g)$Y having the X and Y co-ordinates of each node}
}
\description{
This function checks whether the graph passsed as an argument has X and Y attributes
}
\examples{
spat_degree_change_graph()
}
\keyword{:}
\keyword{attributes}
\keyword{check}

