% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_closeness_change_graph.R
\name{spat_closeness_change_graph}
\alias{spat_closeness_change_graph}
\title{Graph Closeness Rank change}
\usage{
spat_closeness_change_graph(graph, transparency = TRUE, from_alpha = 0,
  to_alpha = 1, step = 0.05)
}
\arguments{
\item{graph}{igraph object with V(g)$X and V(g)$Y having the X and Y co-ordinates of each node}

\item{from_alpha}{specifies start alpha value of spatial closeness calculation. Defaults to 0}

\item{to_alpha}{specifies end alpha value of spatial closeness calculation. Defaults to 1}

\item{step}{species the interval at which the alpha values are to be calculated. Defaults to 0.05}

\item{tranparency}{specifies whether lines in the graph are highlighted according to the maximum rank changes. Defaults to TRUE}
}
\description{
This function creates the graph of spatial closeness rank change between alpha values. Seems to not work if modying only from_alpha or to_alpha
}
\examples{
spat_closeness_change_graph()
}
\keyword{betweenness}
\keyword{graph}
\keyword{spatial}

