% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_degree.R
\name{spat_degree}
\alias{spat_degree}
\title{Calculate Spatial Degree}
\usage{
spat_degree(graph, alpha, rank = TRUE)
}
\arguments{
\item{graph}{igraph object with V(g)$X and V(g)$Y having the X and Y co-ordinates of each node.}

\item{alpha}{specifies alpha value at which degree is to be calculted}

\item{rank}{decides whether to return rank of nodes of raw values. Defaults to TRUE meaning ranks are to be calculated}
}
\description{
This function allows you to express your love of Hello World
}
\examples{

}
\keyword{:}
\keyword{clustering}
\keyword{spatial}

