% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_degree_change_graph.R
\name{spat_degree_change_graph}
\alias{spat_degree_change_graph}
\title{A trial Function}
\usage{
spat_degree_change_graph(graph, transparency = TRUE, from_alpha = 0,
  to_alpha = 1)
}
\arguments{
\item{:}{igraph object with V(g)$X and V(g)$Y having the X and Y co-ordinates of each node
tranparency specifies whether lines in the graph are highlighted according to the maximum rank changes
from_alpha specifies start alpha value of spatial degree calculation
to_alpha specifies end alpha value of spatial degree calculation}
}
\description{
This function allows you to express your love of Hello World
}
\examples{
spat_degree_change_graph()
}
\keyword{:}
\keyword{degree}
\keyword{graph}
\keyword{spatial}

