% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_connection_distance_distribution.R
\name{check_connection_distance_distribution}
\alias{check_connection_distance_distribution}
\title{This function creates a density plot for checking distance of connections}
\usage{
check_connection_distance_distribution(graph, bins = 30)
}
\arguments{
\item{graph}{igraph object with V(g)$X and V(g)$Y having the X and Y co-ordinates of each node}

\item{bins}{number of bins for the histogram. Defaults to 30}
}
\description{
This function creates a density plot for checking distance of connections
}
\keyword{:}
\keyword{clustering}
\keyword{spatial}

