% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spatial_clustering.R
\name{check_spatial_clustering}
\alias{check_spatial_clustering}
\title{A trial Function}
\usage{
check_spatial_clustering(graph)
}
\arguments{
\item{graph}{igraph object with V(g)$X and V(g)$Y having the X and Y co-ordinates of each node}
}
\description{
This function creates a graph for iterative Ripley's K function to check for spatial clusetring of nodes
}
\examples{
check_spatial_clustering()
}
\keyword{:}
\keyword{clustering}
\keyword{spatial}

