% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_shp_SpatNet.R
\name{export_shp_SpatNet}
\alias{export_shp_SpatNet}
\title{Export spatial social network as a point and a line shapefiles in a folder called shp inside the working directory. The shapefiles are not projected.}
\usage{
export_shp_SpatNet(graph, projstr = CRS("+init=epsg:3395"))
}
\arguments{
\item{graph}{igraph object with V(g)$X and V(g)$Y having the X and Y co-ordinates of each node}

\item{projstr}{projection string for the input co-ordinates. Defaults to CRS("+init=epsg:3395").}
}
\description{
This function exports the nodes and edges making up the spatial social network as a point and a line shapefile respectively
}
\examples{

}
\keyword{:}
\keyword{sociogram}
\keyword{spatial}

