% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_degree.R
\name{spat_degree}
\alias{spat_degree}
\title{A trial Function}
\usage{
spat_degree(graph, alpha, rank = TRUE)
}
\arguments{
\item{:}{igraph object with V(g)$X and V(g)$Y having the X and Y co-ordinates of each node
alpha at which the calculation is to be done
rank specifies whether the ranks of the nodes relativr at the specified alpha is to be returned instead of the their raw spatial degree value. Defaults to TRUE}
}
\description{
This function allows you to express your love of Hello World
}
\examples{
spat_degree()
}
\keyword{:}
\keyword{degree}
\keyword{spatial}

