% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_viz.R
\name{structure_viz}
\alias{structure_viz}
\title{A trial Function}
\usage{
structure_viz(graph, resolution = FALSE, num = FALSE)
}
\arguments{
\item{:}{igraph object with V(g)$X and V(g)$Y having the X and Y co-ordinates of each node.}

\item{:}{resolution specifies the grouping of points for numbering. Defaults to NA and numbering cannot be performed}

\item{:}{num specifies whether points should be aggregated and numbered. Requires resolution to be specified}
}
\description{
This function allows you to express your love of Hello World
}
\examples{

}
\keyword{:}
\keyword{clustering}
\keyword{spatial}

